Move tests from short translation units into their logical parents
